further implementation of add_direntry(), still incomplete.

+++ src/fileops.c
@@ -102,7 +102,7 @@ internal_readdir(
 			mino->mino_number);
 	}
 
-	return rd/DIRENTRY_LEN;
+	return (int)rd/DIRENTRY_LEN;
 }
 
 static int
